% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autospill-package.R
\docType{package}
\name{autospill-package}
\alias{autospill}
\alias{autospill-package}
\title{autospill: AutoSpill: A method for calculating spillover coefficients to compensate or unmix high-parameter flow cytometry data}
\description{
To learn more about how to use autospill, please start by browsing the vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavidRach/autospill}
  \item \url{https://davidrach.github.io/autospill/}
  \item Report bugs at \url{https://github.com/carlosproca/autospill/issues}
}

}
\author{
\strong{Maintainer}: Carlos P. Roca \email{carlosproca@gmail.com} (\href{https://orcid.org/0000-0003-0230-1926}{ORCID})

}
\keyword{internal}
