% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_flow_data.r
\name{gate.flow.data}
\alias{gate.flow.data}
\title{Gate events in flow data}
\usage{
gate.flow.data(flow.control, asp, BPPARAM = bpparam())
}
\arguments{
\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}

\item{BPPARAM}{An optional \code{BiocParallelParam} instance determining
the parallel back-end to be used during evaluation, or a list of
\code{BiocParallelParam} instances, to be applied in sequence for nested
calls to \code{BiocParallel} functions.}
}
\value{
List of vectors, one per sample, with the ids of gated events.
}
\description{
Gates events in flow data, based on scatter parameters, for calculation of
compensation.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{read.flow.control}} and
\code{\link{get.autospill.param}}.
}
