% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospill_param.r
\name{get.autospill.param}
\alias{get.autospill.param}
\title{Get AutoSpill parameters}
\usage{
get.autospill.param(param.set = "minimal")
}
\arguments{
\item{param.set}{Character string with the name of a parameter set. Posible
values are:
\describe{
\item{\code{"minimal"}}{No figures or tables.}
\item{\code{"final.step"}}{With figures and tables at final step.}
\item{\code{"paper"}}{With all figures and tables used in AutoSpill
paper.}
\item{\code{"website"}}{With all figures and tables used in AutoSpill
website.}
}}
}
\value{
List of AutoSpill parameters. As reference, see source file
\file{get_autospill_param_minimal.r}
}
\description{
Returns parameters for running a calculation of compensation with AutoSpill.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
