% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_spillover_posnegpop.r
\name{get.marker.spillover.posnegpop}
\alias{get.marker.spillover.posnegpop}
\title{Get spillover coefficients using positive and negative populations}
\usage{
get.marker.spillover.posnegpop(flow.gate, flow.control, asp)
}
\arguments{
\item{flow.gate}{List of vectors with ids of gated events per sample.}

\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
Matrix with spillover coefficients.
}
\description{
Calculates spillover coefficients using positive and negative populations.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{get.marker.spillover}}, \code{\link{gate.flow.data}},
\code{\link{read.flow.control}}, and \code{\link{get.autospill.param}}.
}
