% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_spillover.r
\name{get.marker.spillover}
\alias{get.marker.spillover}
\title{Get spillover coefficients}
\usage{
get.marker.spillover(scale.untransformed, flow.gate, flow.control, asp)
}
\arguments{
\item{scale.untransformed}{Logical indicating if calculation has to be run
in untransformed or transformed (bi-exponential) scale.}

\item{flow.gate}{List of vectors with ids of gated events per sample.}

\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List of two matrices, with regressions intercepts and coefficients.
}
\description{
Calculates spillover coefficients with robust linear models.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{gate.flow.data}}, \code{\link{read.flow.control}}, and
\code{\link{get.autospill.param}}.
}
