% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result_together.r
\name{plot_result.together}
\alias{plot_result.together}
\title{Plot slope error and skewness results together}
\usage{
plot_result.together(flow.control, asp)
}
\arguments{
\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
\code{NULL}.
}
\description{
Plots results of slope error and skewness, together for initial and final
steps of AutoSpill, and calculation with positive and negative populations.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{read.flow.control}} and
\code{\link{get.autospill.param}}.
}
