% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flow_control.r
\name{read.flow.control}
\alias{read.flow.control}
\title{Read set of flow controls}
\usage{
read.flow.control(control.dir, control.def.file, asp)
}
\arguments{
\item{control.dir}{Character string with the directory with the set of
single-color controls.}

\item{control.def.file}{Character string with the CSV file defining the
names and channels of the single-color controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List with the following elements:
\itemize{
\item{antigen}
\item{autof.marker.idx}
\item{event}
\item{event.n}
\item{event.number.width}
\item{event.sample}
\item{expr.data.max}
\item{expr.data.max.ceil}
\item{expr.data.min}
\item{expr.data.tran}
\item{expr.data.untr}
\item{figure.scatter.dir}
\item{flow.set}
\item{gate.parameter}
\item{marker}
\item{marker.n}
\item{marker.original}
\item{sample}
\item{scatter.and.marker}
\item{scatter.and.marker.label}
\item{scatter.and.marker.original}
\item{scatter.parameter}
\item{transform}
\item{transform.inv}
\item{wavelength}
}
}
\description{
Read data and metadata from a set of single-color controls.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{get.autospill.param}}.
}
