% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_marker.r
\name{read.marker}
\alias{read.marker}
\title{Read markers in a set of flow controls}
\usage{
read.marker(control.dir, control.def.file, asp)
}
\arguments{
\item{control.dir}{Character string with the directory with the set of
single-color controls.}

\item{control.def.file}{Character string with the CSV file defining the
names and channels of the single-color controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
Dataframe with the original and corrected names of markers.
}
\description{
Returns a dataframe and writes a csv file with the common set of markers in
a set of single-color controls, together with corrected names in case of
presence of forbidden characters.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{get.autospill.param}}.
}
