% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_spillover.r
\name{refine.spillover}
\alias{refine.spillover}
\title{Refine spillover coefficients}
\usage{
refine.spillover(
  marker.spillover.unco.untr,
  marker.spillover.unco.tran,
  flow.gate,
  flow.control,
  asp
)
}
\arguments{
\item{marker.spillover.unco.untr}{List of two matrices, with regressions
intercepts and coefficients, resulting from the initial spillover
calculation in untransformed scale.}

\item{marker.spillover.unco.tran}{List of two matrices, with regressions
intercepts and coefficients, resulting from the initial spillover
calculation in transformed scale. Optional parameter used only in
scatter plots, it can be \code{NULL}.}

\item{flow.gate}{List of vectors with ids of gated events per sample.}

\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List with four elements:
\itemize{
\item{Spillover matrix at final step.}
\item{Compensation matrix at final step.}
\item{Compensation error at final step, a list with four matrices:
intercepts, coefficients, slopes, and skewness.}
\item{Dataframe with convergence data.}
}
}
\description{
Refines spillover coefficients iteratively.
}
\references{
Roca \emph{et al}:
AutoSpill: A method for calculating spillover coefficients to compensate
or unmix high-parameter flow cytometry data.
\emph{bioRxiv} 2020.06.29.177196;
\href{https://doi.org/10.1101/2020.06.29.177196}{doi:10.1101/2020.06.29.177196}
(2020).
}
\seealso{
\code{\link{get.marker.spillover}}, \code{\link{gate.flow.data}},
\code{\link{read.flow.control}}, and \code{\link{get.autospill.param}}.
}
