% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_flow_data.r
\name{gate.flow.data}
\alias{gate.flow.data}
\title{Gate events in flow data}
\usage{
gate.flow.data(flow.control, asp)
}
\arguments{
\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List of vectors, one per sample, with the ids of gated events.
}
\description{
Gates events in flow data, based on scatter parameters, for calculation of
compensation.
}
\examples{

library(autospill)
Folder <- system.file("extdata",package="autospill")
FolderPath <- file.path(Folder, "MM1")
MetadataPath <- list.files(FolderPath, pattern=".csv", full.names=TRUE)
outpath <- tempdir()
asp <- get.autospill.param("paper", outpath=outpath)
flow.control <- read.flow.control(control.dir=FolderPath,
 control.def.file=MetadataPath, asp=asp)
flow.gate <- gate.flow.data(flow.control=flow.control, asp=asp)

}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{read.flow.control}} and
\code{\link{get.autospill.param}}.
}
