% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autospill_param.r
\name{get.autospill.param}
\alias{get.autospill.param}
\title{Get AutoSpill parameters}
\usage{
get.autospill.param(param.set = "minimal")
}
\arguments{
\item{param.set}{Character string with the name of a parameter set. Posible
values are:
\describe{
\item{\code{"minimal"}}{No figures or tables.}
\item{\code{"final.step"}}{With figures and tables at final step.}
\item{\code{"paper"}}{With all figures and tables used in AutoSpill
paper.}
\item{\code{"website"}}{With all figures and tables used in AutoSpill
website.}
}}
}
\value{
List of AutoSpill parameters. As reference, see source file
\file{get_autospill_param_minimal.r}
}
\description{
Returns parameters for running a calculation of compensation with AutoSpill.
}
\examples{

asp <- get.autospill.param("paper")

}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
