% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_spillover.r
\name{get.marker.spillover}
\alias{get.marker.spillover}
\title{Get spillover coefficients}
\usage{
get.marker.spillover(scale.untransformed, flow.gate = NULL, flow.control, asp)
}
\arguments{
\item{scale.untransformed}{Logical indicating if calculation has to be run in
untransformed or transformed (bi-exponential) scale.}

\item{flow.gate}{List of vectors with ids of gated events per sample, can be
missing if no gating is required.}

\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List of two matrices, with regressions intercepts and coefficients.
}
\description{
Calculates spillover coefficients with robust linear models.
}
\examples{
A <- 2 + 2
}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{gate.flow.data}}, \code{\link{read.flow.control}}, and
\code{\link{get.autospill.param}}.
}
