% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.r
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot AutoSpill convergence}
\usage{
plot_convergence(convergence.log, popnegpop.slope, asp)
}
\arguments{
\item{convergence.log}{Dataframe with convergence data of AutoSpill.}

\item{popnegpop.slope}{Matrix with slope errors resulting from calculating
spillover with positive and negative populations. Optional parameter, it
can be \code{NULL}.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
\code{NULL}.
}
\description{
Plots convergence of iterative refinement of the spillover matrix.
}
\examples{
A <- 2 + 2
}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{refine.spillover}}, \code{\link{process.posnegpop}},
and \code{\link{get.autospill.param}}.
}
