% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result_together.r
\name{plot_result.together}
\alias{plot_result.together}
\title{Plot slope error and skewness results together}
\usage{
plot_result.together(flow.control, asp)
}
\arguments{
\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
\code{NULL}.
}
\description{
Plots results of slope error and skewness, together for initial and final
steps of AutoSpill, and calculation with positive and negative populations.
}
\examples{
A <- 2+2
}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{read.flow.control}} and
\code{\link{get.autospill.param}}.
}
