% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_posnegpop.r
\name{process.posnegpop}
\alias{process.posnegpop}
\title{Process data with positive and negative populations}
\usage{
process.posnegpop(marker.spillover.unco.untr, flow.gate, flow.control, asp)
}
\arguments{
\item{marker.spillover.unco.untr}{List of two matrices, with regressions
intercepts and coefficients, resulting from the initial spillover
calculation in untransformed scale.}

\item{flow.gate}{List of vectors with ids of gated events per sample.}

\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List with three elements:
\itemize{
\item{Calculated spillover matrix.}
\item{Calculated compensation matrix.}
\item{Resulting compensation error, a list with four matrices:
intercepts, coefficients, slopes, and skewness.}
}
}
\description{
Calculates spillover coefficients using positive and negative populations,
and the corresponding compensation error.
}
\examples{
A <- 2+2
}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{get.marker.spillover}}, \code{\link{gate.flow.data}},
\code{\link{read.flow.control}}, and \code{\link{get.autospill.param}}.
}
