% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flow_control.r
\name{read.flow.control}
\alias{read.flow.control}
\title{Read set of flow controls}
\usage{
read.flow.control(control.dir, control.def.file, asp, outpath = NULL)
}
\arguments{
\item{control.dir}{Character string with the directory with the set of
single-color controls or an object of class \code{cytoset}.}

\item{control.def.file}{Character string with the CSV file defining the
names and channels of the single-color controls.}

\item{asp}{List with AutoSpill parameters.}

\item{outpath}{Default NULL which results in the folders being created in the
current working directory. Provide a file.path if you wish the created
created folders to end up elsewhere.}
}
\value{
List with the following elements:
\itemize{
\item{antigen}
\item{autof.marker.idx}
\item{event}
\item{event.n}
\item{event.number.width}
\item{event.sample}
\item{expr.data.max}
\item{expr.data.max.ceil}
\item{expr.data.min}
\item{expr.data.tran}
\item{expr.data.untr}
\item{figure.scatter.dir}
\item{flow.set}
\item{gate.parameter}
\item{marker}
\item{marker.n}
\item{marker.original}
\item{sample}
\item{scatter.and.marker}
\item{scatter.and.marker.label}
\item{scatter.and.marker.original}
\item{scatter.parameter}
\item{transform}
\item{transform.inv}
\item{wavelength}
}
}
\description{
Read data and metadata from a set of single-color controls.
}
\examples{

Folder <- system.file("extdata",package="autospill")
FolderPath <- file.path(Folder, "MM1")
MetadataPath <- list.files(Folder, pattern=".csv", full.names=TRUE)
asp <- get.autospill.param("paper")
flow.control <- read.flow.control(control.dir=FolderPath,
 control.def.file=MetadataPath, asp=asp)

}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{get.autospill.param}}.
}
