% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_marker.r
\name{read.marker}
\alias{read.marker}
\title{Read markers in a set of flow controls}
\usage{
read.marker(control.dir, control.def.file, asp)
}
\arguments{
\item{control.dir}{Character string with the directory with the set of
single-color controls or an object of class \code{cytoset}.}

\item{control.def.file}{Character string with the CSV file or a
\code{data.frame} defining the names and channels of the single-color
controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
Dataframe with the original and corrected names of markers.
}
\description{
Returns a dataframe and writes a csv file with the common set of markers in
a set of single-color controls, together with corrected names in case of
presence of forbidden characters.
}
\examples{

Folder <- system.file("extdata",package="autospill")
FolderPath <- file.path(Folder, "MM1")
MetadataPath <- list.files(Folder, pattern=".csv", full.names=TRUE)
asp <- get.autospill.param("paper")
flow.set.marker.table <- read.marker(control.dir=FolderPath,
control.def.file=MetadataPath, asp=asp)

}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{get.autospill.param}}.
}
