% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_spillover.r
\name{refine.spillover}
\alias{refine.spillover}
\title{Refine spillover coefficients}
\usage{
refine.spillover(
  marker.spillover.unco.untr,
  marker.spillover.unco.tran,
  flow.gate = NULL,
  flow.control,
  asp
)
}
\arguments{
\item{marker.spillover.unco.untr}{List of two matrices, with regressions
intercepts and coefficients, resulting from the initial spillover
calculation in untransformed scale.}

\item{marker.spillover.unco.tran}{List of two matrices, with regressions
intercepts and coefficients, resulting from the initial spillover
calculation in transformed scale. Optional parameter used only in
scatter plots, it can be \code{NULL}.}

\item{flow.gate}{List of vectors with ids of gated events per sample, can be
NULL if no gating is required.}

\item{flow.control}{List with data and metadata of a set of controls.}

\item{asp}{List with AutoSpill parameters.}
}
\value{
List with four elements: \itemize{ \item{Spillover matrix at final
step.} \item{Compensation matrix at final step.} \item{Compensation error
at final step, a list with four matrices: intercepts, coefficients, slopes,
and skewness.} \item{Dataframe with convergence data.} }
}
\description{
Refines spillover coefficients iteratively.
}
\examples{
A <- 2+2
}
\references{
Roca \emph{et al}:
AutoSpill is a principled framework that simplifies the analysis of multichromatic
flow cytometry data.
\emph{Nature Communications} 12, 2890 (2021);
\href{https://doi.org/10.1038/s41467-021-23126-8}{https://doi.org/10.1038/s41467-021-23126-8}
(2021).
}
\seealso{
\code{\link{get.marker.spillover}}, \code{\link{gate.flow.data}},
\code{\link{read.flow.control}}, and \code{\link{get.autospill.param}}.
}
